tiktok-automation/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
orjson>=3.10.0
loguru>=0.7.2
tabulate>=0.9.0
requests>=2.31.0
PyQt6>=6.6.1
PyQt6-Qt6>=6.9.0
uiautomator2>=3.4.0
adbutils>=2.9.4
numpy>=1.26.4
opencv-python>=4.9.0.80
pillow>=10.3.0
pytesseract>=0.3.10
APScheduler>=3.10.4
pyyaml>=6.0
keyring>=25.2.0
